fix search func
